Change deploy stage to have it's own job and not run unit tests (#1574)

+++ _travis/upload_coverage.sh
@@ -2,6 +2,8 @@
 
 set -exo pipefail
 
-source .tox/${TOXENV}/bin/activate
-pip install codecov
-codecov --env TRAVIS_OS_NAME,TOXENV
+if [[ -e .coverage ]]; then
+    source .tox/${TOXENV}/bin/activate
+    pip install codecov
+    codecov --env TRAVIS_OS_NAME,TOXENV
+fi
